Update to mimemagic 0.3.10
Versions below 0.3.7 were yanked from Ruby Gems because it had a dependency that was GPL licensed when it doesn't have a GPL licence itself.  There is now a dependency on the system having the shared-mime-info package installed, but this is already the case on Ubuntu and can be installed easily via Homebrew on a Mac.