Update rubocop
The main changes here have been associated with the removal of
Openstruct.

This was mostly achieved by converting it to a plain hash with
indifferent access. It doesn't appear that we ever accessed the
keys via a `.key_name` call, so this conversion was pretty
straight-forward, and allowed a bit of simplification.

On niggle we did run into is that although the hash has indifferent
access, key comparisons were no longer matching, as internally keys
are stored as strings. Converting PLACEHOLDERS to an array of
strings resolved this issue.